Add files via upload
issue #192. I saw it was assigned to someone else -- if it is a problem that I submitted this, let me know.

Methods check: 
I filled in "cultured-T-cells,dextramer-sort" in methods.identification field: the PBMCs were short-term expanded ex-vivo in an "antigen-specific manner", before being stained with dextramers and sorted. I therefore also filled in TCL, not PBMC in the meta.tissue field.