Add files via upload
issue #85
the method.frequency field is a bit shaky. The author wrote to me: "In all table, # refer to the clone number while n refer to the number of similar sequences shared by different clones. However, the total number of clones analyzed is not mentioned, so it is difficult to calculate a frequency except if you add all the n for each donor." 
So I indeed summed the n per donor and filled the frequency as n/Sum(n).